requirements.txt
psycopg2
redis
channels
daphne
channels_redis
django-compressor
django-tailwind
django-tailwind[reload]